feat: Actualizando caso de prueba de traducir

+++ caso_traducido1.cpp
@@ -1,19 +1,15 @@
 #include <iostream>
 
 int main() {
-  int x = 10;
-
+  double x = 10;
   if (x > 5) {
     std::cout << "x es mayor que 5" << std::endl;
   }
-
   if (x < 5) {
     std::cout << "x es menor que 5" << std::endl;
   }
-
   if (x == 10) {
     std::cout << "x es igual a 10" << std::endl;
   }
-
   return 0;
 }
